Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,8 +2,8 @@
 
 int main() {
     int a,b,c;
-    int average;
-    scanf("%f%f%f",&a,&b,&c);
+    float average;
+    scanf("%d%d%d",&a,&b,&c);
     average=(a+b+c)/3;
     printf("Average: %.2f",average);
     return 0;
